requirements.txt
dj-database-url==0.4.2
Django==1.10.5
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.3.0
pytz==2016.10
requests==2.13.0
requests-mock==1.3.0
django-jsonview==1.0.0
